# services/qxapi/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3000
info:
  contact:
    name: QXSystem
    url: https://github.com/mindevis/qx-project
  description: |-
    QXSystem REST API. Base path `/api/v1`. Agent WSS: `GET /agent/v1/connect`.
    Interactive docs: `GET /swagger/index.html`
  license:
    name: MIT
  title: QX API
  version: "1.0"
paths:
...
  /servers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List servers
      tags:
      - servers
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Dedicated server
      tags:
      - servers
  /servers/{id}:
    delete:
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete server
      tags:
      - servers
    get:
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Server detail (agent_online, minecraft_running)
      tags:
      - servers
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: `Bearer {access_token}`'
    in: header
    name: Authorization
    type: apiKey
  DeviceAuth:
    description: 'Device JWT from POST /launcher/devices/link. Format: `Bearer {device_token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"